requirements.txt
Django==4.2.7
djangorestframework==3.16.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.5.1
Pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0